Add /view_museum and /view_museum_rank

+++ src/main/java/ru/urfu/museumbot/jpa/service/ReviewService.java
@@ -76,6 +76,7 @@ public class ReviewService {
     public void updateReview(Review review) {
         reviewRepository.save(review);
     }
+
     public List<Review> getAllByUser(Long user){
         return reviewRepository.getAllByUser(user);
     }
@@ -85,7 +86,7 @@ public class ReviewService {
      * @param chatId чат пользователя, который оценивает
      * @param rating оценка от 0 до 10
      */
-    public void rateEvent(Long chatId, double rating) {
+    public void rateEvent(Long chatId, int rating) {
         User user = userRepository.getUserByChatId(chatId);
         Event event = eventRepository.getEventById(user.getReviewingEvent());
         Review review = getReview(user, event);
